+++ src/utils/packageJson.ts
@@ -33,7 +33,7 @@ import { readFileSync } from 'fs';
 // @ts-ignore This will be available when the app uses it.
 // eslint-disable-next-line import/no-unresolved
 import packageJsons from '../../../../package.json';
-import type { PackageJson } from './AppTypes';
+import type { PackageJson } from '../../ipc/MetaFiles';
 
 let packageJson: PackageJson = packageJsons;